celery_worker/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.1
prompt_toolkit==3.0.48
propcache==0.2.1
pymodm==0.4.3
pymongo==3.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
